+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `modalClassNames` and `hasCloseButton` to `Modal`.
 - Fix: Update title style on `StandaloneRange` component.
 - Fix: Remove rounding in `pxToRem` and `stepToRem`.